+++ scripts/bundle-optimizer.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import "./bundle-layout-optimizer.js";